refactor: rename module hlc → clock
The module no longer hosts "just the HLC": with the `Clock` port (#142) it is the
home of the clock abstraction — the port, its production adapter `HlcClock`, the
`ManualClock` test stub and the `Timestamp` value. `clock` names that role; `hlc`
named only the default adapter's algorithm.

Pure rename: `src/hlc.rs` → `src/clock.rs`, every `crate::hlc::` → `crate::clock::`,
`pub mod hlc` → `pub mod clock`. The `Clock` port is now also re-exported at the
crate root (`reconcile::Clock`). `HlcClock` keeps its name (the algorithm is still
an HLC) and the module doc still describes the HLC. No behaviour or wire/on-disk
change.

Stacked on the #142 Clock-port branch (the `clock` name presupposes the port).
